requirements.txt
# Processamento de dados
pandas>=2.0.0
numpy>=1.24.0
pyarrow>=14.0.0

# Banco de dados
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
duckdb>=0.9.0

# Validacao de dados
pydantic>=2.0.0

# Testes
pytest>=7.0.0
pytest-cov>=4.0.0

# Utilitarios
python-dotenv>=1.0.0

# Dashboard
streamlit>=1.30.0
plotly>=5.18.0

# Airflow (instalado separadamente no Docker)
# apache-airflow>=2.8.0